[refactor] 경기도 버스 도착정보 모델을 BusArrival.swift에 병합
- GyeonggiBusArrival.swift 내용을 BusArrival.swift로 이동
- GyeonggiBusArrival.swift 파일 삭제
